+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Commands/Interview/AnswerMultipleOptionsLinkedQuestionCommand.cs
@@ -11,9 +11,9 @@ namespace WB.Core.SharedKernels.DataCollection.Commands.Interview
     [MapsToAggregateRootMethod(typeof(Implementation.Aggregates.Interview), "AnswerMultipleOptionsLinkedQuestion")]
     public class AnswerMultipleOptionsLinkedQuestionCommand: AnswerQuestionCommand
     {
-        public int[][] SelectedPropagationVectors { get; private set; }
+        public decimal[][] SelectedPropagationVectors { get; private set; }
 
-        public AnswerMultipleOptionsLinkedQuestionCommand(Guid interviewId, Guid userId, Guid questionId, int[] rosterVector, DateTime answerTime, int[][] selectedPropagationVectors)
+        public AnswerMultipleOptionsLinkedQuestionCommand(Guid interviewId, Guid userId, Guid questionId, decimal[] rosterVector, DateTime answerTime, decimal[][] selectedPropagationVectors)
             : base(interviewId, userId, questionId, rosterVector, answerTime)
         {
             this.SelectedPropagationVectors = selectedPropagationVectors;
