+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Commands/Interview/AnswerMultipleOptionsQuestionCommand.cs
@@ -9,7 +9,7 @@ namespace WB.Core.SharedKernels.DataCollection.Commands.Interview
     {
         public decimal[] SelectedValues { get; private set; }
 
-        public AnswerMultipleOptionsQuestionCommand(Guid interviewId, Guid userId, Guid questionId, int[] rosterVector, DateTime answerTime, decimal[] selectedValues)
+        public AnswerMultipleOptionsQuestionCommand(Guid interviewId, Guid userId, Guid questionId, decimal[] rosterVector, DateTime answerTime, decimal[] selectedValues)
             : base(interviewId, userId, questionId, rosterVector, answerTime)
         {
             this.SelectedValues = selectedValues;
