+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/GroupEnabled.cs
@@ -5,7 +5,7 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview
 {
     public class GroupEnabled : GroupPassiveEvent
     {
-        public GroupEnabled(Guid groupId, int[] propagationVector)
+        public GroupEnabled(Guid groupId, decimal[] propagationVector)
             : base(groupId, propagationVector) {}
     }
 }
